feat(ui): move promo block to header and simplify footer
