+++ docs/development-log.md
@@ -318,3 +318,8 @@
 ## 2025-09-08
 - Исправлен импорт иконки Telegram в `Footer.jsx` на `Send` из `lucide-react`.
 - Запущены `npm run lint` и `npm test` — оба прошли успешно.
+## 2025-09-09
+- Маркетинговый блок перенесён из футера наверх сайта и выстроен в одну строку.
+- Футер сокращён до основных ссылок и слогана.
+- `npm run lint` и `npm test` прошли без ошибок.
+
